# GameDebugClient/src/api/game.swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Rest Adventure - Game API",
    "version": "0.0.0-alpha.0.233+2a1139010fada7f6864f85e7bd96490563a587b1"
  },
  "servers": [
    {
      "url": "https://localhost:7056"
    }
  ],
  "paths": {
    "/game/team/characters": {
      "post": {
        "tags": [
          "Team"
        ],
        "summary": "Create character",
        "operationId": "Characters_CreateCharacter",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Character"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/team/characters/{characterGuid}": {
      "get": {
        "tags": [
          "Team"
        ],
        "summary": "Get character",
        "operationId": "Characters_GetCharacter",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Character"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Team"
        ],
        "summary": "Delete character",
        "operationId": "Characters_DeleteCharacter",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/team/characters/{characterGuid}/history": {
      "get": {
        "tags": [
          "Team"
        ],
        "summary": "Get character history",
        "operationId": "Characters_SearchCharacterHistory",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "PageNumber",
            "in": "query",
            "description": "The page number\n            ",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "PageSize",
            "in": "query",
            "description": "The page size\n            ",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultOfCharacterHistoryEntry"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/team/characters/{characterGuid}/combats": {
      "get": {
        "tags": [
          "Combats"
        ],
        "summary": "Get combats",
        "operationId": "Combats_GetCombats",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Combat"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Character": {
        "type": "object",
        "description": "Character\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "class",
          "health",
          "maxHealth",
          "progression",
          "location",
          "inventory",
          "jobs"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the character\n            ",
            "format": "guid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the character\n            ",
            "minLength": 1
          },
          "class": {
            "description": "The class of the character\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/CharacterClassMinimal"
              }
            ]
          },
          "health": {
            "type": "integer",
            "description": "The health of the character\n            ",
            "format": "int32"
          },
          "maxHealth": {
            "type": "integer",
            "description": "The health of the character\n            ",
            "format": "int32"
          },
          "progression": {
            "description": "The progression of the character\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/ProgressionBarMinimal"
              }
            ]
          },
          "location": {
            "description": "The current location of the character\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/LocationMinimal"
              }
            ]
          },
          "inventory": {
            "description": "The inventory of the character\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Inventory"
              }
            ]
          },
          "jobs": {
            "type": "array",
            "description": "The jobs of the character\n            ",
            "items": {
              "$ref": "#/components/schemas/JobInstance"
            }
          },
          "ongoingAction": {
            "description": "The interaction being performed by the character\n            ",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Action"
              }
            ]
          },
          "plannedAction": {
            "description": "The action that the character has planned for the next tick\n            ",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Action"
              }
            ]
          }
        }
      },
      "CharacterClassMinimal": {
        "type": "object",
        "description": "Character class (minimal)\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the character class\n            ",
            "format": "guid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the character class\n            ",
            "minLength": 1
          }
        }
      },
      "ProgressionBarMinimal": {
        "type": "object",
        "description": "Progression bar (minimal)\n            ",
        "additionalProperties": false,
        "required": [
          "level",
          "experience",
          "nextLevelExperience"
        ],
        "properties": {
          "level": {
            "type": "integer",
            "description": "The level of progression\n            ",
            "format": "int32"
          },
          "experience": {
            "type": "integer",
            "description": "The experience acquired\n            ",
            "format": "int32"
          },
          "nextLevelExperience": {
            "type": "integer",
            "description": "The experience required to reach next level. The value is null when Level is the max level.\n            ",
            "format": "int32"
          }
        }
      },
      "LocationMinimal": {
        "type": "object",
        "description": "Map location minimal information\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "area",
          "positionX",
          "positionY"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the location\n            ",
            "format": "guid",
            "minLength": 1
          },
          "area": {
            "description": "The area associated with the location\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/MapArea"
              }
            ]
          },
          "positionX": {
            "type": "integer",
            "description": "The X coordinate associated with the location\n            ",
            "format": "int32"
          },
          "positionY": {
            "type": "integer",
            "description": "The Y coordinate associated with the location\n            ",
            "format": "int32"
          }
        }
      },
      "MapArea": {
        "type": "object",
        "description": "Map area\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "level"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the area\n            ",
            "format": "guid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the area\n            ",
            "minLength": 1
          },
          "level": {
            "type": "integer",
            "description": "The level of the area\n            ",
            "format": "int32"
          }
        }
      },
      "Inventory": {
        "type": "object",
        "description": "Inventory\n            ",
        "additionalProperties": false,
        "required": [
          "stacks",
          "weight"
        ],
        "properties": {
          "stacks": {
            "type": "array",
            "description": "The entries of the inventory\n            ",
            "items": {
              "$ref": "#/components/schemas/ItemInstanceStack"
            }
          },
          "weight": {
            "type": "integer",
            "description": "The total weight of the items in the inventory\n            ",
            "format": "int32"
          }
        }
      },
      "ItemInstanceStack": {
        "type": "object",
        "description": "Item stack\n            ",
        "additionalProperties": false,
        "required": [
          "itemInstance",
          "count"
        ],
        "properties": {
          "itemInstance": {
            "description": "The item instance representing this stack\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/ItemInstance"
              }
            ]
          },
          "count": {
            "type": "integer",
            "description": "The number of instances in this stack\n            ",
            "format": "int32"
          }
        }
      },
...
      "JobInstance": {
        "type": "object",
        "description": "Job instance\n            ",
        "additionalProperties": false,
        "required": [
          "job",
          "progression"
        ],
        "properties": {
          "job": {
            "description": "The job that is instantiated\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/JobMinimal"
              }
            ]
          },
          "progression": {
            "description": "The progression of the job\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/ProgressionBarMinimal"
              }
            ]
          }
        }
      },
      "JobMinimal": {
        "type": "object",
        "description": "Job (minimal)\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the job\n            ",
            "format": "guid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the job\n            ",
            "minLength": 1
          }
        }
      },
      "Action": {
        "type": "object",
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "move": "#/components/schemas/MoveAction",
            "harvest": "#/components/schemas/HarvestAction",
            "combat-pve-start": "#/components/schemas/StartPveCombatAction",
            "combat-pve-join": "#/components/schemas/JoinPveCombatAction"
          }
        },
        "description": "An action performed by a character\n            ",
        "additionalProperties": false,
        "required": [
          "name",
          "$type"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the action\n            ",
            "minLength": 1
          },
          "$type": {
            "type": "string"
          }
        }
      },
      "MoveAction": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Action"
          },
          {
            "type": "object",
            "description": "Move action\n            ",
            "additionalProperties": false,
            "required": [
              "location"
            ],
            "properties": {
              "location": {
                "description": "The location to which the character is moving\n            ",
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/LocationMinimal"
                  }
                ]
              }
            }
          }
        ]
      },
      "HarvestAction": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Action"
          },
          {
            "type": "object",
            "description": "Harvest action\n            ",
            "additionalProperties": false,
            "required": [
              "harvest",
              "targetId"
            ],
            "properties": {
              "harvest": {
                "description": "The harvest\n            ",
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/HarvestableEntityHarvestMinimal"
                  }
                ]
              },
              "targetId": {
                "type": "string",
                "description": "The target of the harvest\n            ",
                "format": "guid",
                "minLength": 1
              }
            }
          }
        ]
      },
...
      "StartPveCombatAction": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Action"
          },
          {
            "type": "object",
            "description": "PVE combat action\n            ",
            "additionalProperties": false,
            "required": [
              "monsterGroupId"
            ],
            "properties": {
              "monsterGroupId": {
                "type": "string",
                "description": "The group of monsters\n            ",
                "format": "guid",
                "minLength": 1
              }
            }
          }
        ]
      },
      "JoinPveCombatAction": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Action"
          },
          {
            "type": "object",
            "description": "PVE combat action\n            ",
            "additionalProperties": false,
            "required": [
              "monsterGroupId",
              "combatId"
            ],
            "properties": {
              "monsterGroupId": {
                "type": "string",
                "description": "The group of monsters\n            ",
                "format": "guid",
                "minLength": 1
              },
              "combatId": {
                "type": "string",
                "description": "The combat to join\n            ",
                "format": "guid",
                "minLength": 1
              }
            }
          }
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateCharacterRequest": {
        "type": "object",
        "description": "Character creation options\n            ",
        "additionalProperties": false,
        "required": [
          "name",
          "classId"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the character\n            ",
            "minLength": 1
          },
          "classId": {
            "type": "string",
            "description": "The class of the character\n            ",
            "format": "guid",
            "minLength": 1
          }
        }
      },
      "SearchResultOfCharacterHistoryEntry": {
        "type": "object",
        "description": "Search result\n            ",
        "additionalProperties": false,
        "required": [
          "items",
          "pageNumber",
          "pageSize",
          "totalItemsCount",
          "totalPagesCount"
        ],
        "properties": {
          "items": {
            "type": "array",
            "description": "The items found by the query\n            ",
            "items": {
              "$ref": "#/components/schemas/CharacterHistoryEntry"
            }
          },
          "pageNumber": {
            "type": "integer",
            "description": "The page number corresponding to the results that have been selected\n            ",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "The page size used by the search\n            ",
            "format": "int32"
          },
          "totalItemsCount": {
            "type": "integer",
            "description": "The total number of items matching the query\n            ",
            "format": "int32"
          },
          "totalPagesCount": {
            "type": "integer",
            "description": "The total number of pages\n            ",
            "format": "int32"
          }
        }
      },
      "CharacterHistoryEntry": {
        "type": "object",
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "created": "#/components/schemas/CharacterCreatedHistoryEntry",
            "deleted": "#/components/schemas/CharacterDeletedHistoryEntry",
            "teleported": "#/components/schemas/CharacterTeleporteLocationHistoryEntry",
            "moved": "#/components/schemas/CharacterMoveLocationHistoryEntry",
            "inventory-changed": "#/components/schemas/CharacterInventoryChangedHistoryEntry",
            "job-learned": "#/components/schemas/CharacterLearnedJobHistoryEntry",
            "job-gained-experience": "#/components/schemas/CharacterJobGainedExperienceHistoryEntry",
            "job-leveled-up": "#/components/schemas/CharacterJobLeveledUpHistoryEntry",
            "combat-preparation-started": "#/components/schemas/CharacterStartedCombatPreparationHistoryEntry",
            "combat-started": "#/components/schemas/CharacterCombatStartedHistoryEntry",
            "combat-ended": "#/components/schemas/CharacterCombatEndedHistoryEntry"
          }
        },
        "description": "Character history entry\n            ",
        "additionalProperties": false,
        "required": [
          "tick",
          "$type"
        ],
        "properties": {
          "tick": {
            "type": "integer",
            "description": "The tick at which the event happened\n            ",
            "format": "int64"
          },
          "$type": {
            "type": "string"
          }
        }
      },
      "CharacterCreatedHistoryEntry": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CharacterHistoryEntry"
          },
          {
            "type": "object",
            "description": "Character created history entry\n            ",
            "additionalProperties": false
          }
        ]
      },
      "CharacterDeletedHistoryEntry": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CharacterHistoryEntry"
          },
          {
            "type": "object",
            "description": "Character deleted history entry\n            ",
            "additionalProperties": false
          }
        ]
      },
...
      "CharacterLearnedJobHistoryEntry": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CharacterHistoryEntry"
          },
          {
            "type": "object",
            "description": "Character learned job history entry\n            ",
            "additionalProperties": false,
            "required": [
              "jobId",
              "jobName"
            ],
            "properties": {
              "jobId": {
                "type": "string",
                "description": "The unique ID of the job\n            ",
                "format": "guid",
                "minLength": 1
              },
              "jobName": {
                "type": "string",
                "description": "The name of the job\n            ",
                "minLength": 1
              }
            }
          }
        ]
      },
...
      "Combat": {
        "type": "object",
        "discriminator": {
          "propertyName": "$phase",
          "mapping": {
            "preparation": "#/components/schemas/CombatInPreparation",
            "ongoing": "#/components/schemas/OngoingCombat"
          }
        },
        "description": "Combat base\n            ",
        "x-abstract": true,
        "additionalProperties": false,
        "required": [
          "id",
          "$phase"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the combat instance\n            ",
            "format": "guid",
            "minLength": 1
          },
          "$phase": {
            "type": "string"
          }
        }
      },
      "CombatInPreparation": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Combat"
          },
          {
            "type": "object",
            "description": "Combat in preparation\n            ",
            "additionalProperties": false,
            "required": [
              "attackers",
              "attackersOptions",
              "defenders",
              "defendersOptions"
            ],
            "properties": {
              "attackers": {
                "type": "array",
                "description": "The attackers in the combat instance\n            ",
                "items": {
                  "$ref": "#/components/schemas/EntityMinimal"
                }
              },
              "attackersOptions": {
                "description": "The options of the attacker side\n            ",
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CombatFormationOptions"
                  }
                ]
              },
              "defenders": {
                "type": "array",
                "description": "The defenders in the combat instance\n            ",
                "items": {
                  "$ref": "#/components/schemas/EntityMinimal"
                }
              },
              "defendersOptions": {
                "description": "The options of the defender side\n            ",
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CombatFormationOptions"
                  }
                ]
              }
            }
          }
        ]
      },
...
      "OngoingCombat": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Combat"
          },
          {
            "type": "object",
            "description": "Combat instance\n            ",
            "additionalProperties": false,
            "required": [
              "turn",
              "attackers",
              "defenders"
            ],
            "properties": {
              "turn": {
                "type": "integer",
                "description": "The current turn of the combat instance\n            ",
                "format": "int32"
              },
              "attackers": {
                "type": "array",
                "description": "The attackers in the combat instance\n            ",
                "items": {
                  "$ref": "#/components/schemas/CombatEntity"
                }
              },
              "defenders": {
                "type": "array",
                "description": "The defenders in the combat instance\n            ",
                "items": {
                  "$ref": "#/components/schemas/CombatEntity"
                }
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}